+++ Projet-CY-Wire/progC/avl.h
@@ -5,7 +5,7 @@
 // structure permettant de contenir les informations d'une station
 
 typedef struct Station {
-    unsigned short station_id;           // Identifiant de la station  
+    int station_id;           // Identifiant de la station  
     unsigned long long capacite;       // Capacité en kWh 
     unsigned long long somme_conso;    // Somme des valeurs de consommation 
     struct Station* fg;       // Sous-arbre gauche
@@ -18,7 +18,7 @@ unsigned char hauteur(Station* station);
 Station* rotationGauche(Station* y);
 Station* rotationDroite(Station* x);
 int equilibrerAVL(Station* station);
-Station* inserer(Station* racine, unsigned short id, unsigned long long capacite, unsigned long long somme_conso);
+Station* inserer(Station* racine, int id, unsigned long long capacite, unsigned long long somme_conso);
 void parcourinfixe(Station* racine); // Fonction pour parcourir l'arbre en ordre infixe (gauche -> racine -> droite)
 void libererMemoire(Station* racine);
